Creation of a new retries decorator.
